requirements.txt
pyperclip
pyHook-1.5.1-cp36-cp36m-win_amd64.whl
pydub
gTTS